% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataInput.R
\name{DataInput.txt}
\alias{DataInput.txt}
\title{DataInput.txt}
\usage{
DataInput.txt(path = NULL, numBands = 1, IntensityThreshold = T,
  intMar = 0.97)
}
\arguments{
\item{path}{full path to .txt file.}

\item{numBands}{Number of .txt files to process.}

\item{IntensityThreshold}{Reduces the effect of the cosmic rays by integrating the histogram an thresholding to the intRat integration limit.}

\item{intMar}{Density integration limit.}
}
\value{
a rCRSIObj data object.
}
\description{
Reads a .txt file containing Raman data from Reinshaw devices and convert it to an RamanR data object.
}
