% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotClusterImage.R
\name{PlotClusterImage}
\alias{PlotClusterImage}
\title{PlotClusterImage}
\usage{
PlotClusterImage(rCRSIObj, clusterIndex, interpolate = T)
}
\arguments{
\item{path}{full path to .txt file.}

\item{numBands}{Number of .txt files to process.}
}
\value{
a rCRSIObj data object.
}
\description{
Reads a .txt file containing Raman data from Reinshaw devices and convert it to an RamanR data object.
}
