% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotScanimg.R
\name{PlotScanImage}
\alias{PlotScanImage}
\title{PlotScanImage}
\usage{
PlotScanImage(rCRSIObj, wavelength, numImg = 1, interpolate = T)
}
\arguments{
\item{rCRSIObj}{A rCRSI object.}

\item{wavelength}{Vector containing pairs of indexes that define the area under one or more peaks. (Length must be numImg*2)}

\item{numImg}{Number of images to be plotted.}
}
\description{
Generates and plots the image of the "Are under the curve"(AUC) within the given intervals.
}
