% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataInput.R
\name{rCRMIObj}
\alias{rCRMIObj}
\title{rCRMIObj}
\usage{
rCRMIObj(RamanData, numPixels, numBands = 1, RamanShift, BandsLenght, Coords)
}
\value{
a blank rCRMI data object.
}
\description{
Creates a rCRMI object .
}
