% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smoothing.R
\name{SmoothSpectra}
\alias{SmoothSpectra}
\title{SmoothSpectra}
\usage{
SmoothSpectra(rCRSIObj, method = "normal", OddfiltLength = 7,
  NtoZ = TRUE, BackgroundSubs = TRUE, window = 32)
}
\arguments{
\item{rCRSIObj}{a rCRSIObj.}

\item{OddfiltLength}{Integer. Length of the Savitzky-Golay filter.}

\item{NtoZ}{Boolean. Apply negative-to-zero thresholding during the wavelets preprocessing.}

\item{BackgroundSubs}{Boolean. Apply background substraction using a low frequency filter wavelets processing.}

\item{window}{Integer. Number of added points at the begining and ending of the signal to reduce artifacts from the wavelets transform.}
}
\value{
a rCRSIObj object.
}
\description{
Smooths the data contained in a rCRSIObj using the discrete wavelets transform (DWT).
}
