% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataInput.R
\name{DataInput.txt}
\alias{DataInput.txt}
\title{DataInput.txt}
\usage{
DataInput.txt(numBands = 1)
}
\arguments{
\item{PathtoFile}{full path to .txt file.

Reads a .txt file containing Raman data from Reinshaw devices and convert it to an RamanR data object.}
}
\value{
a RamanR data object.
}
\description{
DataInput.txt
}
