% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smoothing.R
\name{SmoothSpectra}
\alias{SmoothSpectra}
\title{SmoothSpectra}
\usage{
SmoothSpectra(RmnObj, ZroThreshold = TRUE, BackgroundSubs = TRUE, ...)
}
\arguments{
\item{RmnObj}{a RmnObj.}

\item{ZroThreshold}{Boolean. Apply negative-to-zero thresholding during the wavelets preprocessing.}

\item{BackgroundSubs}{Boolean. Apply background substraction using a low frequency filter wavelets processing.}
}
\value{
a RmnObj object.
}
\description{
Smooths the data contained in a RmnObj using the discrete wavelets transform (DWT).
}
